zeaz-ultra-pack/backend/api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose==3.3.0
pydantic==2.6.4
